Refactor search into its own reuse component
