+++ .gitignore
@@ -21,6 +21,7 @@ local.*
 local-*.*
 .env.local
 .env.*.local
+idplist-local.*
 
 # Log files
 npm-debug.log*
